requirements.txt
opencv-python
ultralytics
supabase
python-dotenv
numpy
